requirements.txt
# YOLO Multi-Model Detection System - Complete Dependencies
# For video processing, single image analysis, and comprehensive detection

# Core dependencies for multi-modal detection system
ultralytics>=8.0.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
python-dateutil>=2.8.0

# Video processing
ffmpeg-python>=0.2.0

# Audio processing and moderation
librosa>=0.10.0
pydub>=0.25.0
soundfile>=0.12.0

# Text processing and moderation
pytesseract>=0.3.10
SpeechRecognition>=3.10.0
scikit-learn>=1.3.0
joblib>=1.3.0

# Machine learning
pickle-mixin>=1.0.2

# Optional GPU monitoring
psutil>=5.9.0

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Data processing and analysis
matplotlib>=3.7.0
seaborn>=0.12.0

# File handling
pathlib2>=2.3.7

# JSON handling
jsonschema>=4.17.0

# Logging
colorlog>=6.7.0

# System utilities
click>=8.1.0

# Image processing utilities
scikit-image>=0.21.0

# Additional utilities
requests>=2.31.0
urllib3>=2.0.0

# PyTorch installation (choose one based on your system):
# For CPU only: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
# For CUDA 11.8: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# For CUDA 12.1: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121

# External dependencies (install separately):
# - FFmpeg: Download from https://ffmpeg.org/download.html
# - Tesseract OCR: Download from https://github.com/UB-Mannheim/tesseract/wiki
# - Git: Required for some package installations

# INSTALLATION VERIFICATION:
# Run these commands to verify installation:
# python -c "import ultralytics, cv2, torch, librosa, pytesseract; print('All packages imported successfully')"
# ffmpeg -version
# tesseract --version
